Make contacts form on user-show ajaxed
Users can now add contacts and have them render on the page without
reloading the page. They will not be able to add contacts if they have
more than 5, and receive a message letting them know of this if they
reach the limit. The logic governing the limitation is removed from the
user show template, and occurs in the controller now. This also rewrites
the users controller to make that a user can only look at their own page
and are redirected to it if they attempt to request a user show page for
another user (redirects to login if not logged in). Also, a spanish
value in the EN yaml is corrected to an english calue.